requirements.txt
Flask==3.0.0
Frozen-Flask==0.18
Jinja2==3.1.2
Markdown==3.5.1
python-markdown-math==0.8
python-frontmatter==1.0.0
click==8.1.7
watchdog==3.0.0
